+++ packages/script-loader/test/server.test.ts
@@ -1,9 +1,8 @@
 import { describe, expect, it } from "vitest";
-
 import { createScriptLoader } from "../src";
 
 describe("API doesn't break in SSR", () => {
   it("createScriptLoader() - SSR", () => {
-    expect(createScriptLoader({ src: "url" })).toEqual([undefined, expect.any(Function)]);
+    expect(createScriptLoader({ src: "url" })).toEqual(undefined);
   });
 });